Initial ASGI3 (single-callable) support
