+++ mangum/__init__.py
@@ -3,6 +3,7 @@ import asyncio
 import enum
 import traceback
 import urllib.parse
+from functools import partial
 from dataclasses import dataclass, field
 from typing import Any
 
@@ -16,6 +17,7 @@ class ASGICycleState(enum.Enum):
 class ASGICycle:
 
     scope: dict
+    spec_version: int
     body: bytes = b""
     state: ASGICycleState = ASGICycleState.REQUEST
     app_queue: asyncio.Queue = None
@@ -32,8 +34,13 @@ class ASGICycle:
         loop = asyncio.new_event_loop()
         self.app_queue = asyncio.Queue(loop=loop)
         self.put_message({"type": "http.request", "body": body, "more_body": False})
-        asgi_instance = app(self.scope)
-        asgi_task = loop.create_task(asgi_instance(self.receive, self.send))
+
+        if self.spec_version == 3:
+            asgi_instance = app(self.scope, self.receive, self.send)
+        else:
+            asgi_instance = app(self.scope)(self.receive, self.send)
+
+        asgi_task = loop.create_task(asgi_instance)
         loop.run_until_complete(asgi_task)
         return self.response
 
@@ -104,8 +111,10 @@ class Mangum:
 
     app: Any
     debug: bool = False
+    spec_version: int = 2
 
     def __call__(self, *args, **kwargs) -> Any:
+
         try:
             response = self.asgi(*args, **kwargs)
         except Exception as exc:
@@ -166,7 +175,8 @@ class Mangum:
         elif not isinstance(body, bytes):
             body = body.encode()
 
-        response = ASGICycle(scope, binary=binary)(self.app, body=body)
+        cycle = ASGICycle(scope, spec_version=self.spec_version, binary=binary)
+        response = cycle(self.app, body=body)
         return response
 
     def send_response(self, content: str, status_code: int = 500) -> None:
